alterando permissões

+++ routes/web.php
@@ -40,25 +40,25 @@ Route::group(['middleware' => 'auth'], function () {
 
 
     // Usuário
-    Route::get('/index-usuario', [UsuarioController::class, 'index'])->name('usuario.index');
-    Route::get('/show-usuario/{usuario}', [UsuarioController::class, 'show'])->name('usuario.show');
-    Route::get('/create-usuario', [UsuarioController::class, 'create'])->name('usuario.create');
-    Route::post('/store-usuario', [UsuarioController::class, 'store'])->name('usuario.store');
-    Route::get('/edit-usuario/{usuario}', [UsuarioController::class, 'edit'])->name('usuario.edit');
-    Route::put('/update-usuario/{usuario}', [UsuarioController::class, 'update'])->name('usuario.update');
-    Route::get('/edit-usuario-password/{usuario}', [UsuarioController::class, 'editPassword'])->name('usuario.edit-password');
-    Route::put('/update-usuario-password/{usuario}', [UsuarioController::class, 'updatePassword'])->name('usuario.update-password');
-    Route::delete('/destroy-usuario/{usuario}', [UsuarioController::class, 'destroy'])->name('usuario.destroy');
+    Route::get('/index-usuario', [UsuarioController::class, 'index'])->name('usuario.index')->middleware('permission:index-usuario');
+    Route::get('/show-usuario/{usuario}', [UsuarioController::class, 'show'])->name('usuario.show')->middleware('permission:index-usuario');
+    Route::get('/create-usuario', [UsuarioController::class, 'create'])->name('usuario.create')->middleware('permission:index-usuario');
+    Route::post('/store-usuario', [UsuarioController::class, 'store'])->name('usuario.store')->middleware('permission:index-usuario');
+    Route::get('/edit-usuario/{usuario}', [UsuarioController::class, 'edit'])->name('usuario.edit')->middleware('permission:index-usuario');
+    Route::put('/update-usuario/{usuario}', [UsuarioController::class, 'update'])->name('usuario.update')->middleware('permission:index-usuario');
+    Route::get('/edit-usuario-password/{usuario}', [UsuarioController::class, 'editPassword'])->name('usuario.edit-password')->middleware('permission:index-usuario');
+    Route::put('/update-usuario-password/{usuario}', [UsuarioController::class, 'updatePassword'])->name('usuario.update-password')->middleware('permission:index-usuario');
+    Route::delete('/destroy-usuario/{usuario}', [UsuarioController::class, 'destroy'])->name('usuario.destroy')->middleware('permission:index-usuario');
 
 
     // Empresa
-    Route::get('/index-empresa', [EmpresaController::class, 'index'])->name('empresa.index');
-    Route::get('/show-empresa/{empresa}', [EmpresaController::class, 'show'])->name('empresa.show');
-    Route::get('/create-empresa', [EmpresaController::class, 'create'])->name('empresa.create');
-    Route::post('/store-empresa', [EmpresaController::class, 'store'])->name('empresa.store');
-    Route::get('/edit-empresa/{empresa}', [EmpresaController::class, 'edit'])->name('empresa.edit');
-    Route::put('/update-empresa/{empresa}', [EmpresaController::class, 'update'])->name('empresa.update');
-    Route::delete('/destroy-empresa/{empresa}', [EmpresaController::class, 'destroy'])->name('empresa.destroy');
+    Route::get('/index-empresa', [EmpresaController::class, 'index'])->name('empresa.index')->middleware('permission:index-empresa');
+    Route::get('/show-empresa/{empresa}', [EmpresaController::class, 'show'])->name('empresa.show')->middleware('permission:index-empresa');
+    Route::get('/create-empresa', [EmpresaController::class, 'create'])->name('empresa.create')->middleware('permission:index-empresa');
+    Route::post('/store-empresa', [EmpresaController::class, 'store'])->name('empresa.store')->middleware('permission:index-empresa');
+    Route::get('/edit-empresa/{empresa}', [EmpresaController::class, 'edit'])->name('empresa.edit')->middleware('permission:index-empresa');
+    Route::put('/update-empresa/{empresa}', [EmpresaController::class, 'update'])->name('empresa.update')->middleware('permission:index-empresa');
+    Route::delete('/destroy-empresa/{empresa}', [EmpresaController::class, 'destroy'])->name('empresa.destroy')->middleware('permission:index-empresa');
